requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
bcrypt==5.0.0
python-multipart==0.0.12
httpx==0.27.2
pyotp==2.9.0
qrcode[pil]==7.4.2
anthropic==0.34.2
pypdf==5.0.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
umap-learn==0.5.6
numpy==1.26.4
scikit-learn==1.5.0
python-dateutil==2.9.0
python-docx==1.1.2
